Now using CurrentUserAccessor.kt
Before, the current user was always available through the CurrentUser.kt
singleton.  Now we provide access to it through an object which must
be instantiated.  There is also a fake object, FakeCurrentUserAccessor.kt,
which you can send in as a mock.  By default, the places where CurrentUser
is required will use the real thing - CurrentUserAccessor.kt - see the
constructors for TimeRecordingUtilities.kt and AuthenticationUtilities.kt