It's alive! (broken test)
We now have a running manual browser test to see this thing actually work.

Needed to update the sample html files a bit so they won't cause a request for a stylesheet, and sho they will properly send data as a POST when I push the button on the page.  Next step is to parse the data the browser will send back, and soon after have a skeletal form (spooky) of the entire process of a posting of a time entry.  SO. CLOSE.

Co-authored-by: Byron Katz <byron.katz@coveros.com>
Co-authored-by: Matt Taylor <matthew.taylor@coveros.com>
Co-authored-by: Jenna Charlton <jenna.charlton@coveros.com>