Starting to redesign current user WIP
Here's the story so far...

CurrentUser is where we store the user that is logged in.  However, our previous
design for this was to use a singleton.  Since it's one-and-done, that does
not work well with a potentially multi-threaded or event-driven server
approach.  Now, we could use a ThreadLocal type, but we're not sure we want
to use threading anyway.  So in the meantime, let's just go ahead and
make the current user a simple data class and redesign to accommodate that.

Also - by the way - we probably need to make the database a singleton like
we previously had with current user.  So in a way, CurrentUserAccessor and the
PureMemoryDatabase need to switch their patterns.

Co-authored-by: Byron Katz <byron.katz@coveros.com>
Co-authored-by: Matt Taylor <matthew.taylor@coveros.com>
Co-authored-by: Jenna Charlton <jenna.charlton@coveros.com>