Creating some mock objects for testing - WIP
This required putting interfaces on some classes and then doing some redesign.

Adding some rudimentary mapping between files requested and the way we will handle the action in a POST.

Renaming IOHolder to SocketWrapper.kt

In the main function, we create the top-level business classes - those with a suffix of Utilities.  This enables us to need less setup when testing.