Creating some mock objects for testing
This required putting interfaces on some classes and then doing some redesign.

Adding some rudimentary mapping between files requested and the way we will handle the action in a POST.

Renaming IOHolder to SocketWrapper.kt

In the main function, we create the top-level business classes - those with a suffix of Utilities.  This enables us to need less setup when testing.

New capability to store session information

Sessions are data that indicates a user is authenticated.

For example, if a session exists like this:
"abc123", "byron"

It means that byron is authenticated and has abc123 as his
session value.  If there isn't any entry for byron, then he
isn't authenticated.  There can only be zero or one sessions
for a particular user.  Anything else is invalid.