New capability to store session information
Sessions are data that indicates a user is authenticated.

For example, if a session exists like this:
"abc123", "byron"

It means that byron is authenticated and has abc123 as his
session value.  If there isn't any entry for byron, then he
isn't authenticated.  There can only be zero or one sessions
for a particular user.  Anything else is invalid.