CRLF and removing test
Apparently the spec for HTTP requires CR LF - carriage return + line feed - for the endings of each line.  May as well stick to the rules.

Removing the test about disallowing more than one user to have the same employee id - right now, it's unclear how we want all that to work, so leaving it half-baked is fine.