Extracting UI Tests to new project
Note this is not a gradle subproject.  I originally tried that, but then I was tying the two projects at the hip, and when I ran tests from the top it would run the UI tests in the subproject.

This seems more suitable - to have the UI tests entirely separated into its own project.  It is more amenable to a process where we run the UI tests intentionally and separately.

Also removing Sonarqube stuff and project report.  Wasn't really using them.  When that changes, we can reintroduce them.

Updating Kotlin to 1.4.20

Switching to a more colorful and typical gradle output on the command line.