Better image format, other things
webp is a highly compressed image format that is handled by all major desktop browsers.  Not on mobile, but we probably need a different image on mobile anyway.

Incorporating feedback from Chrome's Lighthouse tool to improve some relatively minor things.

Adding a new "logTrace" call, only currently used for logging what we send to the user - it's way too much data to log ordinarily.

Changing the cache - it was highly aggressive and too-long-lived.  Made it just a minute.  Remember - there are two hard problems in computer science and cache invalidation is one of them.

Slight style improvement for the unauthenticated homepage.

Increasing scope of smoke test.