Creating more surrogates
The idea here is that we get more control over what and how we save to disk if we create these intermediate classes to convert to / from when serializing.  The output can be much smaller / simpler.

It was necessary to make the Hash contructor public so we could do this, but I took care of it by simply adding a note to the constructor that it shouldn't be used.  Simpler is better, usually.

Next step - incorporate these during serialization / deserialization and remove @Serializable from all the other stuff through the codebase.