Redesign for maintenance.
Pulling UI tests back into main project
Removing @Serializable from everything that doesn't require it.
Trying to handle tests that manipulate the real on-disk database better - incorporating code to cleanup
Now when we start the server, it requires the location of the database on disk, see the constructor for SocketCommunication.