Fixing a bug
Serious bug - when the system loads the files from disk and builds a new object it failed to pass along the database directory.  So the very first time you start the server, it is recording data to disk, but the second time, it is skipping serialization.

Also added to the smoke test to account for this, added some new logging.