Change: logout deletes sessions for a user
Before, when we would logout we would use that particular session token to delete the session.  That meant if I were logged in on two machines, logging out would only do so on one machine.

Now with this change, logging out on any machine will clear all the sessions.

Also, note: the sessions are basically permanent at the current time.  We do set a timestamp for when they got created, but we're not doing anything with that right now.

Maybe in the future we might incorporate some sort of timeout process.  That is, if the user weren't to log in in a long time.  Or maybe we give the user a choice about how long they should remain logged in.  But again, right now, as long as they haven't clicked logout and they still have the cookie, it's good forever.