Various improvements
Needed to emphasize using a forward slash at the end of the directory name, without which, the system will put the files in the top directory with a prefix of whatever that text is.
Renaming a class.
Simple caching for static files
Setting log settings to a singleton: There's just one settings for the logs in the entire application.  This wasn't quite behaving as expected before.
Condensing all UI tests
Adjusting cache details