Improving / simplifying serialization
Removed several regular expressions from the code, which caused the start-up for the program to take less time - now takes about 1/3 of a second for an empty system with a pre-existing database, versus 1/2 a second before.  Each regular expression requires some time at startup to compile, so by having fewer of them, our system starts faster.

Also, we aren't really following JSON anymore.  Our pattern for data is really simple:
{ key: value , key: value }

that is, the key has a colon at the end, and the key + value have spaces around them.  Also, we have curly braces wrapped around the entire set of data, per line, to make it crystal clear what belongs to a single group of data.

Also, adjusted the logging for gradle tests.  Now that we have 279 tests, it doesn't make as much sense to log each one as it runs, it's just spam.  So I changed it to just show the number of passing tests, and shows a failing test if that happens with full stacktrace.