Getting rid of Detekt.
98.7% of the complaints were needless noise.

It told me it thought handleRequestAndRespond was too complex.
Yeah nope.  I understand what I'm doing.

It thought loggingConfigHtml was too long
It is complaining about a template function that generates HTML text dynamically.

It thought readAndDeserializeTimeEntries was too nested and complex.
I could break it up, yeah, but by splitting this essential piece into multiple, it would likely just increase the cognitive load.

It felt that *EVERY SINGLE NUMBER IN THE CODE REQUIRED A NAMED CONSTANT* .... FFS.
like this is wrong, according to Detekt:

    enum class Month(val ord: Int) {
        JAN(1), FEB(2), MAR(3), APR(4), MAY(5), JUN(6),
        JUL(7), AUG(8), SEP(9), OCT(10), NOV(11), DEC(12);

and apparently this is also wrong:

      val beginDate = LocalDate.of(1980, 1, 1).toEpochDay()
      val endDate = LocalDate.of(2200, 1, 1).toEpochDay()

....

Maybe 4 or 5 of the 320 complaints are valid, it's just got way, WAY too many false positives.  Said another style: way too high noise to signal ratio.  This boy is crying wolf in a horrible way.  Sure I could go back and basically do heavy surgery to quiet it the hell up, but I'm not willing at the moment to do that.

They could really take a page from Sonarqube.  The magic of Sonarqube is that they developed a set of rules, the "Sonar way", that is a very valid set.  If one of those rules fires, it's certain to be a legitimate problem.

It's understandable, coming from a nerd, they'd want to highlight every single problem, but it's the same path that other tools like PMD and Findbugs went, and those tools are hateful.