New tests and ideas
Adding a new test in ServerPerformanceTests that does a full integration test / performance test for adding time entries.

It's ignored right now because it did indeed bring up a problem - if you are adding multiple time entries concurrently (using threads), you will see a ConcurrentModificationException thrown because the code to figure out the next index will see the data changing while it's trying to count the size of the data.

So we've found a problem, what to do?

Well, my research indicated some options.  We could put more locks through the code.  But lots of locks is painful, unlike lox.  Another option is to use a data structure built for concurrency - the ConcurrentHashMap.  This seems an exciting possibility.  I've added an experiment-test -  testTryingOutConcurrentHashMap_PERFORMANCE - to see how it works.  The data is "weakly consistent" on reads - it will be correct for that point in time.