Adding ActionQueue for logs and disk
This is a threaded FIFO - First In First Out.  A queue.  As actions take place, they get put on this queue, and they are run as soon as possible, but on their own thread, in their own time.  That way, if the system is running insanely fast, it can offload the slow IO portions off onto their own threads.

Also added a stop() command so that as we are closing up, at shutdown time, we can wait for the queues to finish, to avoid losing data.

This entailed changes in quite a few places.