Refactoring / reorganizing sessions and users
Maybe ... Maybe we can create some generic functional thing in the database that is public, and then everything that wants something from the database has to go through that tiny (but insanely powerful) interface.  Then we wouldn't need so many functions in the PMD itself - really, just the data and its key accoutrements like serialization

Now that we are starting to use the *Persistence classes to handle the details of working with the database, and the database simply holds the data, it is implying a reorganization of the tests and classes.

This is working pretty well I think.