Bye surrogates
When I stepped back and looked at the bigger picture, I realized that the value of surrogates was not what I had thought.  Originally there was some value, because these were the classes that were marked as "Serializable" and all I had to do was convert the big classes to these and I'd have serialization.  Now that it's all small and simple and easy, having a separate class just for serialization seems kind of pointless.  It just takes a while for that realization to set in sometimes.