Correcting performance tests
It turns out I just wasn't measuring the entire data flow.  In fact, on the test at testEnterTimeAPI_PERFORMANCE, the code was balking very early on when it saw that the user executing the time entry wasn't the owner of that time, and thus we were short-circuiting pretty badly.  Now that we have a more realistic setup, the performance is more in line with the best of what I saw at the server level, which means that instead of feeling blue about a high contrast in performance between two layers, I'm pretty happy with the perf.