Simplifying the readme
Thanks to the new Gradle toolchain technology, we don't need to have our users install the JDK, they simply need to have _any_ version of Java installed.  If we tell them to install the regular (non-JDK) version, the installer will automatically add the binaries to the path properly.

If they want to use the JDK, it's all good, but I am prioritizing the easiest option.