Various improvements
New option: turn all logging off

a useful option if we want to test the application without any logging running

Modifying the 500 error to show the stacktrace if it receives no string message.

Adding a default message for when the user exceeds 24 hours on a time entry

The MainTests wasn't deleting the directory before starting, now it is.