Test improvements
Now we annotate the tests with whether they are an integration test or a performance test.  It's possible to just be one or the other, or both.

Especially in the case of an integration test, we want to indicate whether the test uses a port or a directory, and for the future really anything that could cause us to be accidentally tied to other tests.

(We're not really doing anything yet with these annotations, just leaving them until we figure out what we want.  However, using them to simply mark their tests is already somewhat valuable)

As much as possible, we want the highest possible degree of independence for our tests, so they can run in any order and are maximally independent... and therefore can be parallel as much as want.  That being the case, there's no reason why we couldn't have, heck, millions of tests that are split and end up taking hardly any time to finish.

Split the UI tests into sections related to their focus, to break out of the pattern we'd been using of all in one sequence - it surely complicates things, having to remember what the state could be.  For the UI tests, it actually sets a clean slate entirely at the beginning of each test - server, database, and browser.

Configuring gradle Junit to run tests in parallel