Marking tests as BDD, database test improvements
We were also missing viewing the time entries as a user story in our BDD Html files.

New annotation for BDD tests.

Slight improvement to markDone, now we can include a predicate as to whether it is done.

A few places were writing the database to the disk and I hadn't yet marked them as Integration tests.