Fixing index bug
When the system restarts, it reads data from the files.  Before this fix, the system was starting with an improper value for the index after loading the data.  Now, we examine the data, find the current highest index, and set the nextIndex to that, plus one.

If we don't find any items in the files, we just set the nextIndex to its ordinary start, which is 1.

fixing a test