Converting over to new BDD framework
Now we write discreet intermediate files for the BDD scenarios.  In some subsequent step we will write code to assemble these nicely, but having the individual pieces well-formed, that later step should be easy, and anyway in the meantime we can still see our situation pretty well.

This writes to the build/bdd directory, and each user story gets a sub-directory of its own, along with all the scenarios for that story from wherever they get used.

This simplified things quite a bit when using the framework, also.  There's no need to set up an object before tests, nor any need to run anything after the suite of tests.  Simpler is better.