A highly imperfect fix for ports
In trying to enable more parallel tests, we start up multiple instances of the server simultaneously, each having its own port number.

Unfortunately, Junit4 doesn't _seem_ to give us any ability to have a variable that is above the lifecycle of the entire test suite.  That is, we would like to have a "port" variable that we can increment as we move across all our tests, so that each test gets a unique port.  Instead, we have to assign different ports per class.

For the future: is there some way to do this?  It would certainly help our test maintainability.