Now using read and modify for database
Matt had a great idea here - we need to distinguish between actions on the database where the data is changed, or not.  Instead of having that be all blended in one, we have split it out into two and for the read-only piece, we actually use some code to enforce read-only.

Co-authored-by: Matthew <matthew.taylor@coveros.com>