Switching to ChangeTrackingSet across the board
Now that we have ChangeTrackingSet, we don't need to have different interfaces for data where we save the entire set vs saving just the changed data.  This means we are a step closer to being able to refactor our interfaces to the data elements in PMD - readX and actOnX.  Might be possible to make them generic now for some interface.