Experimenting with new serialization class
This new serialization class for the database, called DatabaseSerialization, encapsulates all the code related to serialization.  This way, our approach to serializing data is exactly the same no matter what the data is.  For example, prior to this, the technique for serializing Time entries was very different than Projects.
This is not yet finished, it's really just begun, but the code we've added is mostly disconnected from the existing code and so everything should still work fine.  Soon we'll want to blend in the new work and that's when the tests are gonna go haywire.
Looking pretty good so far though, we'll very likely be able to use exactly identical code for data access on all types, which means we'll be able to chop out a lot of duplication and also make it much easier to add and manipulate our database going forward.