Moving db persistence to its own file
From now on, the data in the database will follow similar patterns to best enable a general approach.  That is, we want all the data to have a simple integer incrementing index, which we will use as the filename when saving each piece of data.

Yes, each piece of data - a session, a user, a project, and so on - will get its own file.  Each type of data will get its own directory.  Since modern operating systems allow up to 4 billion files per directory, this approach should certainly work for a little while.

Created a new interface, IndexableSerializable, for those types of data that will be written to the disk.

Still need to make sure the loose ends are tied up, since this is a major redesign.

This should make things a little more organized, and making editing and adding data types a little easier and consistent.

Creating a new AbstractDataAccess class which should highly cut back on our boilerplate code for providing access to the data.  Now, writing and reading data should become far more consistent across the board, especially for new data.

Our next steps are to fix the 10 broken tests in PureMemoryDatabaseTests.kt, which test out what happens when something is corrupted in the data.

Also, currently the code for serialization and deserialization is mostly manual and thus error-prone.  Next steps there are to see if we can somehow extract that to a generalized framework.

Test fix for Read_CorruptedData_Projects_BadData.

add instructions for code-with-me

Remove obsolete test and modify test to match changed database behavior

updated assertion message to be more insightful

fixed remaining persistence tests
ALL SHALL PASS! NONE SHALL FAIL!

latest performance results after fixing tests

Co-authored-by: Mitch Hardee <mitchell.hardee@coveros.com>
Co-authored-by: Rich Mills <rich.mills@coveros.com>
Co-authored-by: Matthew <matthew.taylor@coveros.com>
Co-authored-by: Byron Katz <byron.katz@coveros.com>