Stronger typing for serialization / deserialization
Before this, it was a bit sloppier.  When serializing, we would indicate the key as a string, and when deserializing we would select the correct value from the regular expression matches purely by the correct index.  Yuck.

Now, *marginally* better.  We have a convention in place to use an enum that represents the key text, and are able to use that for both the serialization and deserialization side.

Had to add a bit of code to properly extract the keys and values from the regular expression during deserialization.  Had to also change the regular expression to get at the key value properly.

Still, I wonder if there is some way to avoid using the strings at all and just use enums as keys on both sides.  The only complicating factor is that when we deserialize, it's not immediately obvious how we could generically convert strings to the enum for keys from a particular type.  For example, Project has a set of enums for keys: ID and NAME, and Employee has the same keys, but it's technically a different type because it's a public internal class for each.  To make that happen safely I would need some way to properly convert strings, without using reflection.  Maybe something will come to mind.

Also, deserialization is in a private class internal to each type, with an appropriate interface defining its contract, and which takes values in its constructor that are needed during the deserialization process.  Now the deserialize method only requires the string it will process.

In the meantime, this is, as I mentioned, a little better than before.