Maintenance improvements, better logging
Several of the logs that definitely would have been caused by a user weren't mentioning the user.  Now they do.  In making that change, added the ability to set the current user on the TimeEntryPersistence.kt.

Also, noticed a ConcurrentModificationException during AuthenticationPersistenceTests.testCorruptingSessionDataWithMultiThreading.  Changed the set of modified data in the ChangeTrackingSet to be a concurrent set.  Made ConcurrentSet a mutable collection, which entailed implementing some of its interface, but not all.