Simpler buttons, accessibility, adding style
Instead of having links and buttons on the view time page, we can have just buttons, which simplifies our life a bit.

Dabbling with some minor accessibility improvements to see what we like and don't.

In particular, put aria-invalid="false" in a few places so screen readers don't start off telling people "this field is already invalid".  I mean, mentioning it is required is good, and we'll only complain that it's invalid through the default (and highly accessible) validation messages built-in.

Applying style to the logout page, and differentiating between style for the homepage and the auth-homepage