Can now submit and unsubmit, all tests passing
We used to require that the inputs to an endpoint exactly matched the "requiredInputs" element.  This commit relaxes that requirement.  Now we only need to have those required inputs, but also allow others to appear as well.  This was useful so we could send the current time period along, if the user is not on the default view (today's time period)

Added an "unsubmit" mechanism.

On the viewtime page, failing tests were leading us to realize we needed to account for starting out on a different time period when editing and when saving.  That is now corrected.

The submit button turns to an unsubmit if that time period is submitted

The navigation for time periods is still a bit ugly!  We'll soon correct that.

Co-authored-by: Matt Taylor <matthew.taylor@coveros.com>
Co-authored-by: Mitch Hardee <mitchell.hardee@coveros.com>