Redesign
- extracting code for static files
- moving into companion object
- redesign of system startup

There was a problem with logs before this.

The logging mechamism is entirely procedural, and the state is just public.  This means that when I have multiple systems running in parallel during testing, I'll run into issues where one system is shutting down (and stopping the log queue) while another one is sending log entries to that queue... boom.  A better way is to wire the logging into the same state we're building up during system startup.

Converting logs

Switching all the code to use a logging instance passed around.

In the tests I often use a testLogger, but for those tests that actually use a FullSystem to start the server, I often use the logger that is attached to FullSystem.

Creating new interface for logger