Fixing a bug, making a correction
The bug was a nasty thing.

We recently added a timeout to the socket - if we wait 10 seconds listening for the client to say something and they don't, we close the socket, which throws a SocketTimeoutException.  That's all well and good.  But...

it was getting handled by "handleRequest", which was catching *all* exceptions and guess what it tried doing? Yeah, it tried sending something to the user... through the socket ... which was now closed.

(returnData, line 164 in ServerUtilities)

Since the socket was closed, we found ourselves in a weird state trying to send data on a socket that was definitely closed.

Fixed by allowing the exception to propagate one more level higher, into processConnectedClient, and just waltzing out of the keep-alive loop.

Also, a correction: in TEP, we were using actOn when we should have used read - actOn is only worthwhile when we are making a change to data, read is for just reading.