Various improvements (#84)
Fixing bugs and new test categories

Fixing a bug, making a correction

The bug was a nasty thing.

We recently added a timeout to the socket - if we wait 10 seconds listening for the client to say something and they don't, we close the socket, which throws a SocketTimeoutException.  That's all well and good.  But...

it was getting handled by "handleRequest", which was catching *all* exceptions and guess what it tried doing? Yeah, it tried sending something to the user... through the socket ... which was now closed.

(returnData, line 164 in ServerUtilities)

Since the socket was closed, we found ourselves in a weird state trying to send data on a socket that was definitely closed.

Fixed by allowing the exception to propagate one more level higher, into processConnectedClient, and just waltzing out of the keep-alive loop.

Also, a correction: in TEP, we were using actOn when we should have used read - actOn is only worthwhile when we are making a change to data, read is for just reading.

* Small correction

* Several small improvements

Now we can choose which of several sets of tests to run:

gradlew unittest           - runs just the unit tests
"   fasttest - all the fast tests
"   uitest                 - the ui tests
"   integrationtest        - the integration tests
"   performancetest        - the performance tests
"  apitest  - the API tests
"   nouitest  - everything but the ui
gradlew test runs everything, like before

Small functionality change to system - when you add a time entry on a date from a different time period, when it redirects, you will get taken to that time period.

Also, enabled the smoke test for regular runs

Removed a couple of "role" attributes, weren't giving us anything good.