Fixing a bug in nextIndex
It occurred to me that we weren't regularly thoroughly testing the database restore process.  Our UI tests had been working in memory-only mode. I changed it to use a directory, and also to compare the before-restore and after-restore, and then I found a bug.

See, when you delete *all* the data in a particular type, for example sessions or submitted time periods, and then restart the system, the next index to be used was previously calculated by looking at the largest existing id and then adding one.  But what happens if there aren't any?  It sets the next index to merely 1.  This means that the database could have a nextIndex of 5 in memory but when restored has 1.

Now, every time data is persisted to disk, we also record the next index to its own file, and use that when restoring.

In order to find all this out, I added some code for better equality testing in a couple places as well.  When we compare big things like ConcurrentSet or ConcurrentLinkedQueue, we don't want to be distracted by the infinite details inside... just the main pieces.  Yes, there are complications involved - checking equality on those items is only _weakly_ consistent - it can be in the middle of changing and so on - but on the other hand we only use equality checking on tests anyway.  Hopefully :/

Well, that's the nature of code.  Horribly complex.  That's why we test the ever-lasting-heck outta it with automated tests.