Refactoring and redesign
It started so innocently...

I wanted to enable the server to either be provided system properties for the keystore and keystorePass, or to default to using an internal keystore/pass so "it just works".  Mainly so we can test easier.

That went great! But then I realized that unless we are on Java 11, we don't get access to TLS 1.3, the latest and greatest TLS, so of course I changed that in the build.gradle, which enabled me to change GeneralUtilities.kt.

Which led to further refactoring - I didn't need to have multiple logImperatives in the system.  Also, I noticed how there was some duplication in the types (e.g. Employee, Project), so I put my shoulder to the task of handling some of that stuff more generically.  I succeeded, it's true, but my god - just look at DatabaseDiskPersistence.deserialize and shudder.  Shudder.