considering socket leaks
A way to hopefully mostly ensure that we aren't just adding sockets infinitely is to wrap the only area where we use sockets with a try finally, and put socket.close in the finally.  Also, adding trace logging to examine what the current count of sockets is.

Furthermore, ran some local usage tests and jmeter performance tests to ensure this really worked as expected.  The jmeter perf tests just did some simple things again and again.

"Seems to work! Ship it."