Simplifying next index, fixing a bug
Before we had a complex mechamism for recording the next index to disk.  Now we simplify -

as we add more ros in memory, the next index goes up, but if somehow we delete all the rows, it just sets the next index back to 1, in memory.  That way, we are the same on disk and memory, always.