Add new exception for roles, refactor inheritence of CurrentUser. broken tests
Something needs to bubble up when a role tries to do something it cannot do.
Additionally, CurrentUser was wrapping User, changed it to be an is a relation; this avoids very cumbersome currentUser.user accessing, and caused a VERY messy refactor.
Man tests relying on the omniscience of SYSTEM_USER needed to instead use DEFAULT_ADMIN_USER.