Isolating the database code
Making PMD non-open

In trying to make the database extractable, some fundamental changes were needed.

For example, it doesn't make sense to have code _in_ the database
responsible for adding a new employee - that's not really a concern of
the database, it's a concern of the system.  Moved that to the
FullSystem class.

Also, the database no longer has a whole lot of sets as its constructor.
It has a map where you add as many different kinds of data as you might
like.  This means that designating the schema, and starting up the
database, is different.

I also realized I was recreating the data access class every single call
to the database, so that was an easy fix - see for example, lines 12 and
13 in AuthenticationPersistence.

Removing a performance test

techempower does perf tests for various web frameworks.  I had added one
of their specifications into the system to see how we fared.  Not
particularly well - only a thousand requests per second.  Hardly
anything to write home about.  If we want to get more serious, we'll
want to use real profiling tools.

Removing it because it was distracting while I am in the middle of
extracting the database from the system.

removing Jenkinsfile - haven't been using it