removing maintest, test name fixes, gradle fixes
fixes to some test names

removing some unused stuff from gradle - dependency check, sonar
adding file-system watching, suggested by gradle
Applying parallelization to all tests in gradle

Removing maintests: This is the kind of test that should be run from outside the system.  Will need to think how best to do that, but in the meantime, it's acting up and getting worse, so just removing it.