Handle forbidden at AuthUtilities
Before, we were allowing all unauthorized exceptions to be handled at the top level - the ServerUtilities layer.  Now we handle that a bit further down the stack, closer to where it will likely be thrown.  We can still handle anything thrown from further below, but since all business action _should_ be called through code in AuthUtilities, this is a more suitable place to handle it.  It also means that when we test API's, we can assert that we receive the proper status code rather than a thrown exception.