invitations and password for new admin
Now, when the system starts on an empty database, a new random
password is created for the admin and stored in admin_acct.txt

invitations are a thing where, after an admin creates an employee, we create one of these "invitations" to allow a person to register a user and connect it to the employee.

Also doing a bit of maintenance redesign.  Probably doesn't make that much sense to include identifiers like EmployeeId in places like SubmittedPeriod or User - why not just use the real thing, Employee?

New option to set host (aka domain)