Adjusting login, register, create employee
Updating Style and substance.  There doesn't need to be an obvious link to the register page, since that is only for users who have a link (an employee has been created for them already).  Updating the style on login and register.

Also, there doesn't need to be separate pages for creating employees and viewing them.  Putting those together.