Adjusting the standard header and homepage behavior
For a regular user, there's only one action - entering time.  For the admin, there are several actions available.  Adjusting various pages to be simpler - the unauthenticated homepage now redirects to login.  There is no need for a homepage link for non-admins.  The authenticated homepage for a non-admin redirects to time entry.