Including a cleaner edit ability
Before this, if we were on a mobile device and tried to edit a time entry, it was painful, what with all the teensy buttons.  Now there is a big data entry just like when adding a new time entry.  Did some CSS media-query magic to make the regular create and edit disappear.  It should be possible to use either the narrow or wide UI versions, just depending on whether you narrow the browser or not.