Style and HTML adjustments
Adjusting so that the project and the details now resize to fit their content.

Allowing larger project names

Adjust text wrapping - This way it doesn't allow the text out of the box

Required min/max on employee and project

Sorting time entries in reverse.  This way the most recent time entries are closest to the top.

Adding a script to scroll down to the edited row when in desktop mode