Now persistence system configuration
It irked me that every reload would reset the logging config.  Now, there is a new SystemConfiguration which we'll use to store all run-time configurable elements, starting with logging.

In the thick of making this change, I noticed that major errors that occurred in the ActionQueue would stop the queue without a peep.  Looking further, I realized that exceptions really don't get propagated until the future.get() call, and since ActionQueue runs continuously, the exception just killed it quietly.  Changed submit() to execute() and now we see error messages in the logs if the action queue dies.

Also, in the process of this change, made several changes to the ILogger and friends - we're no longer using a map for the various log types, now we use SystemConfiguration.LogSettings

There was also a bit of confusion because the SystemConfiguration will only ever have one entry in the database, which means the nextIndex and id are a bit more hardcoded and static than usual.  Hopefully that doesn't come back to bite me (famous last words)