Wiring ExecutorService throughout
Previously, there were several executor services scattered through the codebase -
one (single threaded) for controlling the server,
one (single threaded) for the ssl server,
one (pooled / cached) for the threads in the servers (to handle client requests) and
one (single threaded) in the actionQueue.

Furthermore, there was a bit of confusion about where the system was actually blocking, waiting for all those threads to complete.

This change places a new "root" executor service at the Main method, and passes it throughout the system, for everything to share.

This may help us, in the future, if there is a possibility to switch the system to a "virtual" thread paradigm, as Project Loom is working towards.  In that case, we might be able to switch the whole system to virtual threads with a single line change.

I also removed the shutdown and await commands in FullSystem.shutdown.  These didn't seem to actually be helping in any way ... maybe just obsolete?  Removing them didn't break any tests and was seemingly just holding things up.

Project loom details:
See
http://cr.openjdk.java.net/~rpressler/loom/loom/sol1_part1.html#threads-are-what-its-all-about

https://hugmanrique.me/posts/testing-java-ea-builds-gradle/

https://wiki.openjdk.java.net/display/loom/Main

I did try to incorporate Project Loom for a bit, but I don't think it's even ready for us to experiment with, at least, not on the main branch.