Stylistic changes for smaller widths
Created a new common width.  Currently, the widths for
pages are differentiated as:

below 400 pixels
401 - 750 pixels
above 751 pixels

This way we can modify style for each level without concern about breaking something we don't see.  It *does* unfortunately mean that any UI changes have to be repeated across sizes, but there's no getting away from that.

I've had the unfortunate experience of situations where the team tried to DRY their CSS.  That is, they tried their hardest to avoid duplication, so they had an intricate style sheet which applied to multiple pages and widths in an impossibly complex way.  The only way to tell of a breakage was to try it out for every width, on every page.

I have found that of the two options, I prefer it this way very much - to have distinct style sheets for particular pages and widths.

To be clear, the two options are:
a) common style patterns are applied across the system, to avoid repetition.... but any changes have to be carefully reviewed on every page (and remember, some pages might be hard to test)
b) recreate the pattern on each page, for each width... but the process of updating style must be carried out on each style sheet, there is tremendous repetition (but any changes you make are only scoped to that page and width)