Minor improvements
Noticed that some elements in the development handbook weren't rendering properly.  fixed.

Adjusted the style on the create project and create employee pages to take full width.

On the create employee, the invitation is now just text, not an anchor tag (which might cause people to click it, which is pointless)